+++ spec/database_helper.rb
@@ -1,13 +1,12 @@
 
-def clear_arcade
-  #  delete testdatabase and restore it
+def  connect
+  #  create testdatabase and connect
   databases =  Arcade::Api.databases
   if databases.nil?
   puts "Edit Credentials in config.yml"
   Kernel.exit
   end
-  if  databases.include?(Arcade::Config.database[:test])
-    Arcade::Api.drop_database Arcade::Config.database[:test] 
-  end
-  Arcade::Api.create_database Arcade::Config.database[:test]
+    Arcade::Api.create_database Arcade::Config.database[:test] if  !databases.include?(Arcade::Config.database[:test])
+    Arcade::Init.connect :test
+
 end